requirements.txt
fastapi
uvicorn
twilio
openai
python-decouple
python-multipart
pyngrok